Remove int16 dtype specifications in piano roll matrices
This was creating issues downstream due to recent numpy changes which
did not allow implicit downcasting of float-dtype matrices when they
operated with integer matrices (see #58).  In retrospect, there is no
real reason for these matrices to be of int16 dtype, so we will just use
the numpy default of float64 everywhere.