Create separate insruments for different tracks/channels
Before this change, when loading in a MIDI file, a single Instrument
instance is created for all events on all channels and all tracks which
have the same program number.  Beyond being potentially an unwanted
collapsing in some applications, it also results in incorrect behavior
due to pitch/control messages being applied to all instruments with the
same program number in a given MIDI file.  This change adds
functionality for keeping track of instruments present on each
track/channel/program number.

Resolves #60.