Test for time signature after adjusting for tempo
If a beat's location is adjusted because the prior innter-beat-interval
crosses a tempo change boundary, it's possible that it will have the
same location as a time signature change event, which is not caught when
the time signature change test appears above the tempo correction.  This
puts the time signature change test below, which fixes this.