Fix initial tempo/time signature index search
The initial tempo change index and time signature change index should
correspond to the first which appears before start_time.  As previously
implemented, it was the first equal to or after start_time, which was
generaly OK as start_time was generally 0 and there is always a tempo
change at time 0, but when start_time > 0 and/or there is no time
signature change at time zero, this was wrong.  This fixes it.