Add adjust_events subfunction
The procedure for adjusting pitch bends and control changes is basically
the same, and requires some additional special logic which was not there
before for adjusting and removing events outside of the valid range.
This adds a subfunction which includes all of the functionality, and
calls it twice for pitch bends and control changes.