Raise when requiring IOIs with too few notes
Resolves #36.

When estimating a single global tempo, there needs to be at least two
notes so that at least one inner-onset-interval can be computed.  This
raises a ValueError whenever estimate_tempo is called when there are
fewer than two notes.  Also, in order to estimate the beat start, there
needs to be at least one note, so we raise a ValueError there too.  We
don't need to raise a ValueError in estimate_tempo because it just
returns empty np.ndarrays.