Use searchsorted on self.__tick_to_time for time_to_tick
Resolves #73

This speeds up time_to_tick, up to a factor of about 300x depending on
how many entries there are in self._tick_scales.  Some special logic is
required to get the index of the closest element, but this method (with
nearest_index) is still about 4x faster than np.argmin(np.abs)).