Move past time signature if ==
Before, if start_time was equal to a time signature change time, it
we would not move past that time signature because
`beats[-1] !< time_signature.time`.  This adds an `=` so that we move
past the time signature if it has the same time as the beat.  This makes
things work in the case where `start_time` is a time signature change
time.