Update current BPM at the end of each loop iteration
With get_current_bpm at the top of this loop iteration, the conditional
was being tested on the old BPM and the new tempo change time.  This
probably wasn't a big deal when the tempo was not changing a lot, but
when it does change a lot, this would create the wrong behavior.  It
should be at the bottom, after the tempo index is updated.  And, with it
here, we don't need to udpate the tempo after we exit the loop.