Fix integrated piano roll division by zero
When supplying a time array to piano roll, if the time array
exceeds the length of the "native" piano roll generated based
on when the last note finishes, the mean of empty arrays is
computed, resulting in warnings and populating the outputted
piano roll with nan. Similarly, if the time array has a greater
sampling rate than the "native" piano roll, more empty arrays
are generated, resulting in nan when numpy.mean is computed.

- Fixed the above issues as long as time array is increasing
  and greater than or equal to 0.
- Added unit test for the fix
- Made casting to integer explicit in instrument.py to stop
  the numpy deprecation warning